feat: cleaned up file structures
