--- apps/DDA-server/src/file/dto/create-file.input.ts
@@ -1,1 +1,0 @@
-export class CreateFileInput {}
