--- apps/DDA-server/src/file/dto/update-file.input.ts
@@ -1,6 +1,0 @@
-import { CreateFileInput } from './create-file.input';
-import { PartialType } from '@nestjs/mapped-types';
-
-export class UpdateFileInput extends PartialType(CreateFileInput) {
-  id: number;
-}
