+++ apps/DDA-server/src/file/file.service.ts
@@ -1,12 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { CreateFileInput } from './dto/create-file.input';
-import { UpdateFileInput } from './dto/update-file.input';
 
 @Injectable()
 export class FileService {
-  create(createFileInput: CreateFileInput) {
-    return 'This action adds a new file';
-  }
 
   findAll() {
     return `This action returns all file`;
@@ -14,10 +9,6 @@ export class FileService {
 
   findOne(id: number) {
     return `This action returns a #${id} file`;
-  }
-
-  update(id: number, updateFileInput: UpdateFileInput) {
-    return `This action updates a #${id} file`;
   }
 
   remove(id: number) {
