--- apps/DDA-server/src/file_tags/dto/create-file_tag.input.ts
@@ -1,1 +1,0 @@
-export class CreateFileTagInput {}
