--- apps/DDA-server/src/file_tags/dto/update-file_tag.input.ts
@@ -1,6 +1,0 @@
-import { CreateFileTagInput } from './create-file_tag.input';
-import { PartialType } from '@nestjs/mapped-types';
-
-export class UpdateFileTagInput extends PartialType(CreateFileTagInput) {
-  id: number;
-}
