+++ apps/DDA-server/src/file_tags/file_tags.resolver.ts
@@ -1,7 +1,5 @@
 import { Resolver, Query, Mutation, Args } from '@nestjs/graphql';
 import { FileTagsService } from './file_tags.service';
-import { CreateFileTagInput } from './dto/create-file_tag.input';
-import { UpdateFileTagInput } from './dto/update-file_tag.input';
 
 @Resolver('FileTag')
 export class FileTagsResolver {
