--- apps/DDA-server/src/tag/dto/create-tag.input.ts
@@ -1,1 +1,0 @@
-export class CreateTagInput {}
