--- apps/DDA-server/src/tag/dto/update-tag.input.ts
@@ -1,6 +1,0 @@
-import { CreateTagInput } from './create-tag.input';
-import { PartialType } from '@nestjs/mapped-types';
-
-export class UpdateTagInput extends PartialType(CreateTagInput) {
-  id: number;
-}
