+++ apps/DDA-server/src/tag/tag.service.ts
@@ -1,23 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import { CreateTagInput } from './dto/create-tag.input';
-import { UpdateTagInput } from './dto/update-tag.input';
 
 @Injectable()
 export class TagService {
-  create(createTagInput: CreateTagInput) {
-    return 'This action adds a new tag';
-  }
-
   findAll() {
     return `This action returns all tag`;
   }
 
   findOne(id: number) {
     return `This action returns a #${id} tag`;
-  }
-
-  update(id: number, updateTagInput: UpdateTagInput) {
-    return `This action updates a #${id} tag`;
   }
 
   remove(id: number) {
