--- apps/DDA-server/src/user/dto/create-user.input.ts
@@ -1,1 +1,0 @@
-export class CreateUserInput {}
