--- apps/DDA-server/src/user/dto/update-user.input.ts
@@ -1,6 +1,0 @@
-import { CreateUserInput } from './create-user.input';
-import { PartialType } from '@nestjs/mapped-types';
-
-export class UpdateUserInput extends PartialType(CreateUserInput) {
-  id: number;
-}
