--- apps/DDA-server/src/user/entities/user.entity.ts
@@ -1,17 +1,0 @@
-import { FileEntity } from 'src/file/entities/file.entity';
-import { Entity, Column, PrimaryGeneratedColumn, OneToMany } from 'typeorm';
-
-@Entity("user")
-export class UserEntity {
-    @PrimaryGeneratedColumn()
-    id: number;
-
-    @Column()
-    username: string;
-
-    @Column()
-    password: string;
-
-    @OneToMany(() => FileEntity, file => file.user)
-    files: FileEntity[];
-}
